Replace qr.js with new library, package directly
Note: the QR Codes generated are not identical. This seems to be due to
the exact approach taken when encoding strings (leaving encoding to the
third-party lib again). If there are issues with Kanji or other
multibyte strings, we'll need to revisit this decision, potentially fork
this new lib or use it differently.

https://www.nayuki.io/page/qr-code-generator-library