Rewrite Canvas renderer with hooks
Much like the SVG renderer, this gets rid of PureComponent and classes
and goes to function components. This shouldn't matter in practice. Part
of the reasoning was that getDerivedStateFromProps didn't give us what
I needed, so we could get rid of state and simplify the logic. Now this
actually checks the image element's status to determine if its loaded.

In the process here, I was able to fix a bug to actually handle when
images fail to load and excavation is set. Previously it would leave
a whole in the middle, now it doesn't.