Add new `marginSize` prop
This allows specifying the margin size (in modules) rather than just
allowing 0 or 4. While 4 is still encouraged to comply with spec, this
allows people to do what they need.

For the next major version, `includeMargin` will continue to work, but
`marginSize` will override and `includeMargin` will be removed for the
following major version.